+++ src/main/java/br/pucrio/les/esg_token_backend/services/value/IValueService.java
@@ -8,4 +8,5 @@ public interface IValueService {
 
     public List<Value> index();
 
+    public Value create(Value value);
 }